# packages/shared-types/src/index.ts
export type UserRole = 'user' | 'guide' | 'admin'

export interface User {
  id: string
  email: string
  name: string
  role: UserRole
}

export interface Trail {
  id: string
  name: string
  state: string
  difficulty: 'easy' | 'moderate' | 'hard'
  distanceKm: number
  durationMinutes: number
  hasWater?: boolean
}

export interface TripPlan {
  id: string
  title: string
  userId: string
  startDate: string
  endDate: string
  itinerary: Array<{ day: number; trailId: string; notes: string }>
  checklist: string[]
}